shardeum-hub/requirements.txt
Django==4.0
django-ckeditor==5.4.0
django-cleanup==6.0.0
django-crispy-forms==1.7.2
django-js-asset==1.0.0
